Generalize animation and use it for other things
