+++ web/components/VenueNames/VenueNames.module.css
@@ -4,6 +4,20 @@
   padding: 0;
 }
 
+.venuesEnter .venue {
+  transform-origin: bottom;
+  animation: item-enter var(--duration) cubic-bezier(0.25, 0.46, 0.45, 0.94)
+    both;
+  animation-delay: var(--delay);
+}
+
+@keyframes item-enter {
+  from {
+    opacity: 0;
+    transform: translateY(var(--distance)) rotate(var(--rotation));
+  }
+}
+
 .venue {
   display: inline-block;
   font: var(--font-display-base);
